Project is getting ready to be "finnished"
The big things that are missing are to not press on the things that you can unlock. Also to make some levels with some enemies and potions. Then the basic prototype is done. Still a lot of bugs and not a lot of the nice to have would be done. But we can call it done and use that as the final project, then just say "Look how much we have done since, and how good are we yay" Ez 12:d