The normal Particles is almost dont
Need to make the damage numbers, so that the max amount of hits it has taken within a certain time frame is the same as max amount of particles.
This will make it add a new to the max each time you hit it (make a cap on it, so it dosent bug out), and then set the max amount down one after a particle is removed. Should work