Missing some sounds
Have made the distance sound, so we can have something like a torch, or the sounds on the weapons, so we cant hear them from a mile away.
Want to lerp the two sound tracks, change more sounds. Then i will make the boss better, make more particles (maybe for the player weapon, and the spawn)
Want to make the potion drop, but later make it drop something like the blood, that heals.
Make a chest for more coins or a random drop
And make the a boost or eat a mushroom and become a beserker or something. WOuld be fun i quess